+++ src/components/interfaces.ts
@@ -1,3 +1,5 @@
 export interface TestableComponent {
     "data-test-id"?: string
 }
+
+export type { ContextOverlayProps } from "./ContextOverlay/ContextOverlay";